requirements.txt
# Web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
aiofiles==23.2.1

# Pydantic / FastAPI transitive deps
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
anyio==4.10.0
click==8.1.8
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0

# Image / CV
numpy>=1.26.0
pillow>=10.0.0
opencv-python-headless>=4.8.0

# Background removal (U2-Net via rembg)
rembg==2.0.28

# Small utilities
tqdm
scikit-image
scipy
